refactor(telemetry): address PR #63 review comments
- Remove runtime cast of ITelemetryStatistics to TelemetryStatistics in
  DI constructor; store interface + nullable internal reference instead
- Remove unused 'version' parameter from TelemetryBuilder.AddActivitySource()
- Add null guard for Features in AddHttpInstrumentation()
- Move TelemetryOptionsValidator to Configuration namespace (matches folder)
- Remove unused using directives across 6 source/test files
- Replace foreach loop with LINQ .Where().Count() in builder tests
- Simplify discard assignment in StaticTelemetryTests
